add new Page as a Homescreen/Startscreen
